# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "description": "Pose un cookie de session `httpOnly`, valable le temps du jeton.",
        "operationId": "AuthController_connecter",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnexionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JetonDto"
                }
              }
            }
          },
          "401": {
            "description": "Identifiants invalides"
          }
        },
        "summary": "Connexion par matricule et mot de passe",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/moi": {
      "get": {
        "description": "Le front s’en sert au démarrage pour savoir s’il a une session : le cookie étant `httpOnly`, il ne peut pas le lire lui-même.",
        "operationId": "AuthController_moi",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentConnecteDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "summary": "Identité et permissions de l'agent connecté",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/mot-de-passe": {
      "post": {
        "description": "Invalide tous les jetons de l'agent, y compris celui de l'appel : le jeton neuf renvoyé remplace le précédent, et le cookie est reposé.",
        "operationId": "AuthController_changerMotDePasse",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangementMotDePasseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JetonDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "summary": "Changement de mot de passe",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/deconnexion": {
      "post": {
        "description": "Retire le cookie de session. Le jeton reste valide jusqu’à sa péremption — pour le révoquer immédiatement, c’est `token_version` qu’il faut incrémenter, ce que fait un changement de mot de passe.",
        "operationId": "AuthController_deconnecter",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "summary": "Déconnexion",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/dossiers": {
      "get": {
        "description": "Les dossiers privés en sont absents, sans mention. Le décompte des entités suivies ne compte que le visible.",
        "operationId": "DossiersController_lister",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DossierResumeDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "summary": "Liste des dossiers",
        "tags": [
          "dossiers"
        ]
      },
      "post": {
        "description": "Nom, entité pivot, visibilité et note. Le pivot est suivi dès la création.",
        "operationId": "DossiersController_creer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreationDossierDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PanneauDossierDto"
                }
              }
            }
          },
          "409": {
            "description": "Un dossier porte déjà ce nom"
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "summary": "Création d’un dossier",
        "tags": [
          "dossiers"
        ]
      }
    },
    "/dossiers/{id}": {
      "get": {
        "description": "Ouvrir un dossier revient à ouvrir la fiche de son entité pivot ; ce panneau est ce que la fiche affiche en plus lorsqu’on y arrive par le dossier.",
        "operationId": "DossiersController_panneau",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PanneauDossierDto"
                }
              }
            }
          },
          "404": {
            "description": "Inconnu, ou privé sans habilitation"
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "summary": "Panneau de dossier",
        "tags": [
          "dossiers"
        ]
      },
      "patch": {
        "operationId": "DossiersController_modifier",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModificationDossierDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PanneauDossierDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "tags": [
          "dossiers"
        ]
      }
    },
    "/dossiers/{id}/suivi": {
      "post": {
        "description": "Aucune duplication : l’entité reste unique, une même fiche peut être suivie par plusieurs dossiers.",
        "operationId": "DossiersController_suivre",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DesignationEntiteDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "summary": "Ajout d’une entité au suivi",
        "tags": [
          "dossiers"
        ]
      }
    },
    "/dossiers/{id}/suivi/{entiteId}": {
      "delete": {
        "operationId": "DossiersController_nePlusSuivre",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entiteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "409": {
            "description": "L’entité pivot ne se retire pas du suivi"
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "tags": [
          "dossiers"
        ]
      }
    },
    "/dossiers/{id}/habilitations": {
      "post": {
        "description": "L’habilitation est nominative, jamais déduite d’un grade. Elle s’ajoute aux gardiens qu’un agent doit franchir.",
        "operationId": "DossiersController_habiliter",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DesignationAgentDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "summary": "Whitelist du dossier",
        "tags": [
          "dossiers"
        ]
      }
    },
    "/dossiers/{id}/habilitations/{agentId}": {
      "delete": {
        "operationId": "DossiersController_retirerHabilitation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "agentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "tags": [
          "dossiers"
        ]
      }
    },
...
  },
  "info": {
    "title": "Centrale N&I",
    "description": "API du service Narcotics & Investigations de la SAHP. Le référentiel — types d'entités, champs, types de liens, onglets — est dynamique et n'apparaît pas dans ce contrat : il se récupère par GET /referentiel.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "jeton": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ConnexionDto": {
        "type": "object",
        "properties": {
          "matricule": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64,
            "example": "2291"
          },
          "motDePasse": {
            "type": "string",
            "minLength": 1,
            "maxLength": 256
          }
        },
        "required": [
          "matricule",
          "motDePasse"
        ]
      },
      "AgentConnecteDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "matricule": {
            "type": "string"
          },
          "prenom": {
            "type": "string"
          },
          "nom": {
            "type": "string"
          },
          "roleCode": {
            "type": "string"
          },
          "superAdmin": {
            "type": "boolean"
          },
          "doitChangerMdp": {
            "type": "boolean",
            "description": "Tant qu'il vaut true, seules /auth/moi et /auth/mot-de-passe répondent"
          },
          "permissions": {
            "description": "Codes de permissions du grade. Un super-admin contourne ce jeu.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "matricule",
          "prenom",
          "nom",
          "roleCode",
          "superAdmin",
          "doitChangerMdp",
          "permissions"
        ]
      },
      "JetonDto": {
        "type": "object",
        "properties": {
          "jeton": {
            "type": "string",
            "description": "À placer en en-tête Authorization: Bearer …"
          },
          "agent": {
            "$ref": "#/components/schemas/AgentConnecteDto"
          }
        },
        "required": [
          "jeton",
          "agent"
        ]
      },
      "ChangementMotDePasseDto": {
        "type": "object",
        "properties": {
          "ancien": {
            "type": "string",
            "minLength": 1,
            "maxLength": 256
          },
          "nouveau": {
            "type": "string",
            "minLength": 12,
            "maxLength": 256
          }
        },
        "required": [
          "ancien",
          "nouveau"
        ]
      },
...
      "DossierResumeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nom": {
            "type": "string"
          },
          "visibilite": {
            "enum": [
              "public",
              "restreint",
              "prive"
            ],
            "type": "string"
          },
          "etat": {
            "enum": [
              "actif",
              "archive"
            ],
            "type": "string"
          },
          "entitePivotId": {
            "type": "string",
            "format": "uuid"
          },
          "entitePivotLibelle": {
            "type": "string"
          },
          "nombreSuivis": {
            "type": "number",
            "description": "Entités suivies **visibles par cet agent**"
          },
          "creeLe": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "nom",
          "visibilite",
          "etat",
          "entitePivotId",
          "entitePivotLibelle",
          "nombreSuivis",
          "creeLe"
        ]
      },
      "EntiteSuivieDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "libelle": {
            "type": "string"
          },
          "typeCode": {
            "type": "string"
          },
          "estPivot": {
            "type": "boolean"
          },
          "ajouteLe": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "libelle",
          "typeCode",
          "estPivot",
          "ajouteLe"
        ]
      },
      "AgentHabiliteDto": {
        "type": "object",
        "properties": {
          "agentId": {
            "type": "string",
            "format": "uuid"
          },
          "libelle": {
            "type": "string",
            "description": "« agent supprimé » si le compte est anonymisé"
          },
          "matricule": {
            "type": "string"
          },
          "accordeLe": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "agentId",
          "libelle",
          "matricule",
          "accordeLe"
        ]
      },
      "PanneauDossierDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nom": {
            "type": "string"
          },
          "visibilite": {
            "enum": [
              "public",
              "restreint",
              "prive"
            ],
            "type": "string"
          },
          "etat": {
            "enum": [
              "actif",
              "archive"
            ],
            "type": "string"
          },
          "entitePivotId": {
            "type": "string",
            "format": "uuid"
          },
          "entitePivotLibelle": {
            "type": "string"
          },
          "nombreSuivis": {
            "type": "number",
            "description": "Entités suivies **visibles par cet agent**"
          },
          "creeLe": {
            "type": "string",
            "format": "date-time"
          },
          "contenuLisible": {
            "type": "boolean",
            "description": "Faux sur un dossier restreint sans habilitation : le nom s’affiche, le reste non"
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "suivis": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EntiteSuivieDto"
            }
          },
          "habilitations": {
            "description": "Whitelist. Vide tant que le contenu n’est pas lisible.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentHabiliteDto"
            }
          }
        },
        "required": [
          "id",
          "nom",
          "visibilite",
          "etat",
          "entitePivotId",
          "entitePivotLibelle",
          "nombreSuivis",
          "creeLe",
          "contenuLisible",
          "note",
          "suivis",
          "habilitations"
        ]
      },
      "CreationDossierDto": {
        "type": "object",
        "properties": {
          "nom": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128,
            "example": "Madrina"
          },
          "entitePivotId": {
            "type": "string",
            "format": "uuid",
            "description": "Obligatoire : un dossier sans entité pivot n’existe pas. Ouvrir le dossier revient à ouvrir sa fiche."
          },
          "visibilite": {
            "enum": [
              "public",
              "restreint",
              "prive"
            ],
            "type": "string"
          },
          "note": {
            "type": "string",
            "maxLength": 10000,
            "description": "Champ libre, sans source ni fiabilité — ce n’est pas un fait"
          }
        },
        "required": [
          "nom",
          "entitePivotId"
        ]
      },
      "ModificationDossierDto": {
        "type": "object",
        "properties": {
          "nom": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "note": {
            "type": "string",
            "maxLength": 10000
          },
          "visibilite": {
            "enum": [
              "public",
              "restreint",
              "prive"
            ],
            "type": "string",
            "description": "Exige la permission visibilite.definir"
          }
        }
      },
      "DesignationEntiteDto": {
        "type": "object",
        "properties": {
          "entiteId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "entiteId"
        ]
      },
      "DesignationAgentDto": {
        "type": "object",
        "properties": {
          "agentId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "agentId"
        ]
      },
...
    }
  }
}